% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Synthetic dataset of 1000 responses to 10 items, each with four categories, generated from a Partial Credit Model.}
\format{An object of class \code{data.frame} with 1000 rows and 10 columns.}
\usage{
dataset
}
\description{
Synthetic dataset of 1000 responses to 10 items, each with four categories, generated from a Partial Credit Model.
}
\keyword{datasets}
